+++ include/shared/logs.go
